Kenzie/Sub-MAF-21: Moved optimal range listeners into their own folders (not sure why I didn't do that before...)
